% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temporal_BD.r
\name{temporal_BD}
\alias{temporal_BD}
\title{Temporal beta-diversity}
\usage{
temporal_BD(data_in = NULL, time_step_col_name = NULL,
  taxon_list = NULL, taxon_id_col_name = NULL,
  biomass_col_name = NULL)
}
\arguments{
\item{data_in}{Data input for analysis (can be either long or wide format). When \code{data_in} is wide format and rows are times and columns are taxa, no other parameters are requred. See examples for different \code{data_in} format types that are accepted.}

\item{time_step_col_name}{Character string indicating time step column name. For use with both \code{data_long} and \code{data_wide} inputs.}

\item{taxon_list}{Vector (character) of taxon IDs, used to indicate which columns in \code{data_wide} are species counts (or biomass measurements)}

\item{taxon_id_col_name}{Character string indicating column name in \code{data_long} that contains taxon IDs. Required for long format data.}

\item{biomass_col_name}{Character string indicating column name in \code{data_long} that contains biomass measurements (or species counts). Required for long format data.}
}
\value{
Returns a numeric value.
}
\description{
Calculates temporal beta-diversity for a single site
}
\examples{
# data example
library(dplyr)
library(tidyr)

dat_1site_5times <- data.frame(
  a = c(1, 2, 3, 4, 5),
  b = c(23, 5, 76, 4, 56),
  c = c(0, 0, 12, 41, 1))

dat_1site_5times_with_time_col <- data.frame(
  my_time_step = c(1, 2, 3, 4, 5),
  a = c(1, 2, 3, 4, 5),
  b = c(23, 5, 76, 4, 56),
  c = c(0, 0, 12, 41, 1))

dat_long <- dat_1site_5times_with_time_col \%>\%
  gather(my_species, my_biomass, -my_time_step)

# examples
# wide data, only species cols, rows ordered by time
temporal_BD(data_in = dat_1site_5times)

# wide data, indicate species and time_step columns
temporal_BD(data_in = dat_1site_5times_with_time_col,
            taxon_list = c('a', 'b', 'c'), time_step_col_name = 'my_time_step')

# long data
temporal_BD(data_in = dat_long,
            time_step_col_name = 'my_time_step',
            taxon_id_col_name = 'my_species',
            biomass_col_name = 'my_biomass')

}
\author{
Eric R. Sokol \email{sokole@gmail.com}

Thomas Lamy \email{thomas.lamy27@gmail.com}
}
